+++ src/app/app-routing.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
-import { HomeComponent } from './gallery/components/home/home.component'
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
+  { path: '', loadChildren: './gallery/gallery.module#GalleryModule' },
   { path: 'cms', loadChildren: './cms/cms.module#CmsModule' },
 ]
