+++ src/app/cms/components/cms-root/cms.component.ts
@@ -1,6 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import {AuthService} from '../../services/auth.service'
-
+import { Component, OnInit } from '@angular/core'
+import { AuthService } from '../../services/auth.service'
 
 @Component({
   selector: 'app-cms',
