+++ src/app/gallery/components/common/header/header.component.html
