+++ src/app/gallery/components/common/header/header.component.spec.ts
