+++ src/app/gallery/components/gallery-root/gallery.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core'
+import { HttpClient } from '@angular/common/http'
 
 @Component({
   selector: 'app-gallery',
@@ -6,7 +7,15 @@ import { Component, OnInit } from '@angular/core'
   styleUrls: ['./gallery.component.scss'],
 })
 export class GalleryComponent implements OnInit {
-  constructor() {}
+  constructor(private http: HttpClient) {}
 
-  ngOnInit() {}
+  ngOnInit() {
+    console.log('start')
+
+    this.http
+      .get('https://us-central1-analog-pics-a1a3b.cloudfunctions.net/photos')
+      .subscribe(el => {
+        console.log(el)
+      })
+  }
 }
