+++ src/app/gallery/components/routes/home/home.component.spec.ts
@@ -1,19 +1,19 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing'
 
-import { FileUploadComponent } from './file-upload.component'
+import { HomeComponent } from './home.component'
 
-describe('FileUploadComponent', () => {
-  let component: FileUploadComponent
-  let fixture: ComponentFixture<FileUploadComponent>
+describe('HomeComponent', () => {
+  let component: HomeComponent
+  let fixture: ComponentFixture<HomeComponent>
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [FileUploadComponent],
+      declarations: [HomeComponent],
     }).compileComponents()
   }))
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(FileUploadComponent)
+    fixture = TestBed.createComponent(HomeComponent)
     component = fixture.componentInstance
     fixture.detectChanges()
   })
