+++ src/app/gallery/gallery-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
-import { HomeComponent } from './components/home/home.component'
+import { HomeComponent } from './components/routes/home/home.component'
 import { GalleryComponent } from './components/gallery-root/gallery.component'
 
 const routes: Routes = [
